+++ lib/src/json/adapter/list_json_serialization_adapter.dart
@@ -1,3 +1,17 @@
+// ---------------------------------------------------------------------------
+// 🍃 JetLeaf Framework - https://jetleaf.hapnium.com
+//
+// Copyright © 2025 Hapnium & JetLeaf Contributors. All rights reserved.
+//
+// This source file is part of the JetLeaf Framework and is protected
+// under copyright law. You may not copy, modify, or distribute this file
+// except in compliance with the JetLeaf license.
+//
+// For licensing terms, see the LICENSE file in the root of this project.
+// ---------------------------------------------------------------------------
+// 
+// 🔧 Powered by Hapnium — the Dart backend engine 🍃
+
 import 'package:jetleaf_lang/lang.dart';
 
 import '../context/json_deserialization_context.dart';
@@ -130,9 +144,11 @@ final class ListJsonSerializationAdapter extends JsonSerializationAdapter<List> 
   @override
   void serialize(List value, JsonGenerator generator, JsonSerializationContext serializer) {
     generator.writeStartArray();
+    
     for (final item in value) {
       serializer.serialize(item, generator);
     }
+
     generator.writeEndArray();
   }